requirements.txt
virtualenv==20.16.7
pillow==9.2.0
numpy==1.23.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66